backend/pyproject.toml
[project]
name = "wairz-backend"
version = "0.1.0"
description = "Wairz - AI-Assisted Firmware Reverse Engineering & Security Assessment"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "python-magic>=0.4.27",
    "pyelftools>=0.31",
    "aiofiles>=24.0.0",
    "weasyprint>=62.0",
    "pypdf>=4.0",
    "packageurl-python>=0.15.0",
    "nvdlib>=0.7.6",
    "docker>=7.0.0",
    "httpx>=0.27.0",
    "mcp[cli]>=1.2.0",
    "cryptography>=43.0.0",
    "yara-python>=4.5.0",
    "redis>=5.0.0",
    "arq>=0.26.0",
    "lief>=0.15.0",
    "pefile>=2024.8",
    "capstone>=5.0.0",
    "aiosqlite>=0.20.0",
    "rapidfuzz>=3.0.0",
    "scapy>=2.6",
    "bandit>=1.7",
    "clamd>=1.0.2",
    "androguard>=4.1.2",
    "mobsfscan>=0.3.0",
    "semgrep>=1.80.0",
    "setuptools>=70.0",
]

[project.scripts]
wairz-mcp = "app.mcp_server:main"
wairz-scan = "app.cli.scan:main"
wairz-compare-apk = "app.cli.compare_apk:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
